requirements.txt
yfinance
pandas
python-telegram-bot==20.3
ta
tabulate
flask